claude-code-tools/doc-rag/requirements.txt
chromadb>=1.0.0
fastapi>=0.115.0
uvicorn>=0.34.0
mcp>=1.0.0
anthropic>=0.52.0
openai>=1.30.0
python-dotenv>=1.0.0
